Undo the package merging of files
