+++ src/main/java/Duke/Duke.java
@@ -2,13 +2,14 @@ package Duke;
 
 import Duke.Exception.EmptyArgumentException;
 import Duke.Exception.WrongArgumentException;
-import Duke.Tasks.Task;
+import Duke.Tasks.TaskType;
+import Duke.Tasks.TaskList;
 
 import java.util.Scanner;
 
 
 public class Duke {
-    private static Task.TaskList list = new Task.TaskList();
+    private static TaskList list = new TaskList();
 
     private static final String SEPARATOR = "____________________________________________________________";
 
@@ -137,7 +138,7 @@ public class Duke {
         if (lineInput.strip().isEmpty()) {
             throw new EmptyArgumentException();
         }
-        int index = list.addTaskToList(lineInput, Task.TaskType.TODO);
+        int index = list.addTaskToList(lineInput, TaskType.TODO);
         String output = "I got you, added a todo:\n"
                 + list.getItemFromList(index + 1)
                 + "\n Now you have " + (index + 1) + " tasks in the list.";
@@ -148,7 +149,7 @@ public class Duke {
         if (lineInput.strip().isEmpty()) {
             throw new EmptyArgumentException();
         }
-        int index = list.addTaskToList(lineInput, Task.TaskType.DEADLINE);
+        int index = list.addTaskToList(lineInput, TaskType.DEADLINE);
         String output = "I got you, added a deadline:\n"
                 + list.getItemFromList(index + 1)
                 + "\n Now you have " + (index + 1) + " tasks in the list.";
@@ -159,7 +160,7 @@ public class Duke {
         if (lineInput.strip().isEmpty()) {
             throw new EmptyArgumentException();
         }
-        int index = list.addTaskToList(lineInput, Task.TaskType.EVENT);
+        int index = list.addTaskToList(lineInput, TaskType.EVENT);
         String output = "I got you, added a event:\n"
                 + list.getItemFromList(index + 1)
                 + "\n Now you have " + (index + 1) + " tasks in the list.";
