+++ src/main/java/Duke/Tasks/TaskType.java
@@ -1,0 +1,5 @@
+package Duke.Tasks;
+
+public enum TaskType {
+    TODO, DEADLINE, EVENT, OTHER
+}
